# src/main/java/com/eaglebank/interfaces/rest/dto/request/AuthenticationRequest.java
package com.eaglebank.interfaces.rest.dto.request;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;

/**
 * Request DTO for user authentication
 */
public record AuthenticationRequest(
        @NotBlank(message = "Email is required")
        @Email(message = "Email must be valid")
        String email,

        @NotBlank(message = "Password is required")
        String password
) {
}